requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.8
DateTime==4.3
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
idna==2.10
lxml==4.6.3
mysql-connector-python==8.0.24
numpy==1.21.2
packaging==20.9
pandas==1.3.3
pandas-datareader==0.10.0
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.23
urllib3==1.26.3
zope.interface==5.4.0
